Repaired Linux compilation; fixed a few warnings Had to futz with line endings in game/ai/ai.cpp before SVN would STFU, so unfortunately some false deltas were created.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@1268 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/ai/ai.h
@@ -882,7 +882,7 @@ protected:
 	// Grace period info for each alert level
 	float m_gracetime_1, m_gracetime_2, m_gracetime_3;
 	float m_gracefrac_1, m_gracefrac_2, m_gracefrac_3;
-	float m_gracecount_1, m_gracecount_2, m_gracecount_3;
+	int m_gracecount_1, m_gracecount_2, m_gracecount_3;
 	// De-alert times for each alert level
 	float atime1, atime2, atime3;
